rendering asset rates in Prices component
