+++ styles/index.js
@@ -1,6 +1,6 @@
-import * as page from './page';
-import * as header from './header';
-import * as prices from './prices';
+import * as page from './page'
+import * as header from './header'
+import * as prices from './prices'
 
 export {
   page,
